translated default acq methods on POL view
